requirements.txt
Flask==3.0.0
pymongo[srv]==4.6.1
python-dotenv==1.0.0
flask-bcrypt==1.0.1
